move some py content to sdk
minor update
