+++ .gitignore
@@ -3,3 +3,6 @@ bin/dummy_task
 bin/worker
 .idea
 .DS_Store
+*.whl
+*.egg-info
+sdk/python/dist
