+++ workload/docker/python/pydummy/Dockerfile
@@ -7,6 +7,10 @@ RUN apt-get install -y netcat-openbsd net-tools
 WORKDIR /
 
 # COPY ./scripts/start.sh /start
-COPY ./src/start.py /start
+COPY ./workload/docker/python/pydummy/src/start.py /start
+
+COPY ./sdk/python/dist/spear*.whl /tmp/
+RUN pip install /tmp/spear*.whl
+RUN rm /tmp/spear*.whl
 
 CMD ["sleep", "infinity"]
