+++ workload/docker/python/pydummy/compose.yaml
@@ -2,5 +2,5 @@ services:
   dummy:
     image: pysimple
     build:
-      context: .
-      dockerfile: Dockerfile
+      context: ../../../../
+      dockerfile: ./workload/docker/python/pydummy/Dockerfile
