+++ workload/docker/python/pydummy/src/demo.go
@@ -1,11 +1,8 @@
 package main
 
 import (
-	"bufio"
 	"bytes"
-	"fmt"
 	"net/http"
-	"os"
 
 	"github.com/lfedgeai/spear/pkg/tools/docker"
 	log "github.com/sirupsen/logrus"
@@ -16,15 +13,6 @@ func init() {
 }
 
 func main() {
-	// get input from user
-	reader := bufio.NewReader(os.Stdin)
-	fmt.Print("Image Description: ")
-
-	input, err := reader.ReadString('\n')
-	if err != nil {
-		panic("reader.ReadString failed: " + err.Error())
-	}
-
 	// setup test environment
 	s := docker.NewTestSetup()
 	defer s.TearDown()
@@ -32,7 +20,7 @@ func main() {
 	// send a http request to the server and check the response
 	client := &http.Client{}
 	req, err := http.NewRequest("GET", "http://localhost:8080", bytes.NewBuffer(
-		[]byte(input),
+		[]byte("Hello, World!"),
 	))
 
 	if err != nil {
